requirements.txt
aioconsole==0.8.1
aiohttp==3.12.14
colorama==0.4.6
Pillow~=11.3.0
Requests==2.32.4
rich==13.7.1
urllib3==2.5.0
selenium==4.22.0
webdriver-manager==4.0.2
future~=1.0.0
cryptography
pycryptodome~=3.23.0
pypresence

PyQt5>=5.15.11
WMI>=1.5.1
beautifulsoup4>=4.13.4
textual>=0.10.0